Eliminate dead code for LimsUtils
This also relocates methods that have only one use to the point of use
