+++ core/src/test/java/uk/ac/bbsrc/tgac/miso/core/util/LimsUtilsTest.java
@@ -4,21 +4,10 @@ import static org.junit.Assert.*;
 
 import org.junit.Test;
 
-import uk.ac.bbsrc.tgac.miso.core.data.Sample;
 import uk.ac.bbsrc.tgac.miso.core.data.SampleSlide;
-import uk.ac.bbsrc.tgac.miso.core.data.impl.SampleImpl;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.SampleSlideImpl;
 
 public class LimsUtilsTest {
-
-  @Test
-  public void testValidateRelationshipForSimpleSample() throws Exception {
-    Sample child = new SampleImpl(); // Simple sample has no DetailedSample attributes.
-    Sample parent = null; // Simple sample has no parent.
-    assertTrue(
-        "Simple sample with a null parent and null DetailedSample is a valid relationship",
-        LimsUtils.isValidRelationship(null, parent, child));
-  }
 
   @Test
   public void testInstanceOfSampleTissueProcessing() throws Exception {
