+++ miso-service/src/test/java/uk/ac/bbsrc/tgac/miso/service/impl/DefaultSampleServiceTest.java
@@ -27,18 +27,18 @@ import com.eaglegenomics.simlims.core.User;
 import com.google.common.collect.Lists;
 
 import uk.ac.bbsrc.tgac.miso.core.data.DetailedSample;
-import uk.ac.bbsrc.tgac.miso.core.data.SampleIdentity;
 import uk.ac.bbsrc.tgac.miso.core.data.Project;
 import uk.ac.bbsrc.tgac.miso.core.data.ReferenceGenome;
 import uk.ac.bbsrc.tgac.miso.core.data.Sample;
 import uk.ac.bbsrc.tgac.miso.core.data.SampleClass;
+import uk.ac.bbsrc.tgac.miso.core.data.SampleIdentity;
 import uk.ac.bbsrc.tgac.miso.core.data.SampleStock;
 import uk.ac.bbsrc.tgac.miso.core.data.SampleTissue;
 import uk.ac.bbsrc.tgac.miso.core.data.SampleValidRelationship;
-import uk.ac.bbsrc.tgac.miso.core.data.impl.SampleIdentityImpl;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.ProjectImpl;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.ReferenceGenomeImpl;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.SampleClassImpl;
+import uk.ac.bbsrc.tgac.miso.core.data.impl.SampleIdentityImpl;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.SampleImpl;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.SampleStockImpl;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.SampleTissueImpl;
@@ -640,6 +640,15 @@ public class DefaultSampleServiceTest {
     Mockito.when(authorizationManager.getCurrentUser()).thenReturn(user);
     Mockito.when(securityStore.getUserById(user.getUserId())).thenReturn(user);
     return user;
+  }
+
+  @Test
+  public void testValidateRelationshipForSimpleSample() throws Exception {
+    Sample child = new SampleImpl(); // Simple sample has no DetailedSample attributes.
+    Sample parent = null; // Simple sample has no parent.
+    assertTrue(
+        "Simple sample with a null parent and null DetailedSample is a valid relationship",
+        DefaultSampleService.isValidRelationship(null, parent, child));
   }
 
 }
